+++ packages/react/tests/DirectionProvider.test.tsx
@@ -3,7 +3,7 @@
 import React from 'react';
 import { render } from 'rut-dom';
 import aesthetic from 'aesthetic';
-import { setupAesthetic, teardownAesthetic } from 'aesthetic/lib/testUtils';
+import { setupAesthetic, teardownAesthetic } from 'aesthetic/lib/testing';
 import DirectionProvider from '../src/DirectionProvider';
 import { DirectionProviderProps } from '../src/types';
