+++ packages/react/tests/useTheme.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'rut-dom';
 import aesthetic from 'aesthetic';
-import { setupAesthetic, teardownAesthetic } from 'aesthetic/lib/testUtils';
+import { setupAesthetic, teardownAesthetic } from 'aesthetic/lib/testing';
 import useTheme from '../src/useTheme';
 import ThemeProvider from '../src/ThemeProvider';
